Wrote all OCT/USBL/VFR to db for NA171/H2091
Implemented SQLAlchemy ORM with a local SQLite3 database. Worked
through lots of issues getting this working. The single biggest
problem was unique ID's - I was hoping to use raw_ts as the primary
key, but no dice. I went down a long rabbit hole thinking there was
some problem with values that were None not correctly getting
converted to NULL in SQL world. That was not the case.

The root cause was a combination of using raw_ts as the primary key
just didn't work, and then I had more issues trying to get 'id'
working between the SQLAlchemy ORM model and the actual Pydantic
models.

In Pydantic2, the handling of Optional attributes is much stricter.
Specifically, id: Optional[int] - when the model is init'd, there
must be a key id with *some* value in there. It can be None, or 0,
or whatever, but it needs to exist.

For a new record, there isn't an id yet, and I don't want to do it
in Pydantic. That should be set by the database. Long story short,
I ended up adding a to_orm method on BaseDatum which drops id
from the model dictionary if it's present. That worked and I can
now sucessfully put over a million individual values into the db.

Some other things - this project has a mixture of Pydantic models
and SQLAlchemy ORM models. I added model/lookup.py, which contains
some dicts containing modelname (str, key) to class object mappings.

This is still just the "text" format data, references to images
will still need to come in. However, I think I'm going to work
on reimplementing the actual processing code next, since that's
a more complicated task than implementing more raw datasources.