Initial test read of the HercOCTDatum table.
Simple SQLAlchemy queries are up and running! Right now, all that
I'm doing with the queried data is printing the raw timestamp
to logger.info. However, I was able to work out a few issues
between the Pydantic & SQLAlchemy models.

The to_orm/from_orm process is now bidirectional, and these methods
are defined on the Pydantic BaseDatum model. The to_orm method
just drops the 'id' field if it's present. This is to ensure that
for new values entering the table, that the SQL backend will
automatically assign these, rather than using the default pydantic
value of 0. The from_orm process strips out two things we don't
want ending up in the Pydantic model:

  - "_sa_instance_state", which isn't in the model and is just
    accounting state for SQLAlchemy. It doesn't help us, and it's
    only there since i'm using the __dict__ method.
  - "id". My original plan for all of this was to use raw_ts as
    the main column where I would be selecting specific records,
    however, this plan may change.