Pull HercOCTDatum, decimate to 1Hz, and update HercOCTDatum in the db
Figured out the whole roundtrip workflow between SQLite/ORM/Pydantic.
For OCT & VFR we simply drop messages (at least, for now) to get the
data rate down to 1Hz. This was originally written explicitly in the
target() method of SelectOCTCommand, but I'll pull this out as a
utility method.

The Sonardyne "binning" is more elegant, since there's an explicit
accuracy parameter we can use to select the best points in a 1sec-
long bin.

All SELECTs and UPDATEs are via the bulk API which is good.