Move all SQLAlchemy functions under command/ to BaseCommand
Before, I had some SQL related code split between ImportRawCommand
and SelectOCTCommand. Now, that's all centralized as "private"
methods under BaseCommand. There are five methods that have been
entered in the file, although not all of them are actually
implemented yet.

-check table (create if it doesn't exist)
-check/insert
-read/select
-update
-delete

This follows the standard CRUD model for accessing databaess. The
idea is that as I add specific code for dealing with data-specific
behavior (right now, the standard is that Octans messages are just
decimated to 1Hz, but VFR messages check for fix type and vehicle,
and USBL messsage look for the best accuracy), keeping the individual
command class implementations focused on the actual logic will
prevent this from turning into a rats nest.