Added HercParoDatum (paroscientific depth sensor) & prefixindex
The paroscientific data from Hercules is logged to the .HER files,
along with a ton of other stuff. This is the first bit of data
I've had to pull out of these records, and the prefix attribute
defined in BaseDatum and modified elsewhere didn't work. The issue
is that *every* line in the raw/datalog/*.HER files starts with
"HER". The actual sensor specific prefix doesn't show up until
sline[3].

This is not a super common case, so I've added a prefixindex
attribute on BaseDatum, which defaults to 0 (beginning of the line).
If cls.prefixindex == 0, the prefix test is done with the existing
line.startswith(cls.prefix) test. If it's not 0, then we actually
split the line (on any whitespace, split()) and go from there.
The split is wrapped in a Try/Except IndexError block, so if
prefixindex is wrong, the test just gets '' instead of throwing
an exception. This should result in getting no data, prompting
further investigation.