Fixed invalid USBL & NavEst (VFR) Data
USBL - I forgot to select beacon ID, so I was getting a mixture of
ship, Hercules, and Atalanta positions mixed together.
VFR - The indices in my parser were wrong, so what was picked as
latitude & longitude was totally incorrect (I think it was actually
the GPS Altitude field) which is totally meaningless to us.